requirements.txt
openpyxl
pdfplumber
google-generativeai
python-dotenv
Pillow
